requirements.txt
instagrapi==2.0.0
pandas==2.0.3
gspread==5.12.0
oauth2client==4.1.3
schedule==1.2.0
Pillow==10.0.0
requests==2.31.0
python-dotenv==1.0.0
